pyproject.toml
[project]
name = "cudemodaysautotests"
version = "0.1.0"
description = ""
authors = [
    {name = "NikshetSteh",email = "88205797+NikshetSteh@users.noreply.github.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pytest (>=8.4.1,<9.0.0)",
    "numpy (>=2.3.2,<3.0.0)",
    "python-levenshtein (>=0.27.1,<0.28.0)",
    "pytest-timeout (>=2.4.0,<3.0.0)",
    "pydantic-settings (>=2.10.1,<3.0.0)",
    "requests (>=2.32.4,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
isort = "^6.0.1"
black = "^25.1.0"
pylint = "^3.3.7"

